requirements.txt
flask
google-genai
python-dotenv
flask-sqlalchemy
psycopg2-binary
vercel
fsrs
PyPDF2